Move CPUID related structs and functions into kvm/cpuid.go
